add exception handling

+++ src/main/java/com/collegemanagement/feesmanagement/services/impl/CourseServicesImpl.java
@@ -1,6 +1,7 @@
 package com.collegemanagement.feesmanagement.services.impl;
 
 import com.collegemanagement.feesmanagement.entity.Course;
+import com.collegemanagement.feesmanagement.exception.CourseNotFoundException;
 import com.collegemanagement.feesmanagement.repository.CourseRepository;
 import com.collegemanagement.feesmanagement.services.CourseServices;
 import org.springframework.stereotype.Service;
@@ -28,12 +29,15 @@ public class CourseServicesImpl implements CourseServices {
 
     @Override
     public Course fecthCourseById(Integer id) {
-        return repository.findById(id).get();
+
+        return repository.findById(id).orElseThrow(()->new CourseNotFoundException("course with id "+id+" doesn't exist."));
     }
 
     @Override
     public Course updateCourse(Integer id, Course course) {
-        Course existingCourse = repository.findById(id).get();
+
+        Course existingCourse = repository.findById(id)
+                .orElseThrow(() -> new CourseNotFoundException("Course with id " + id + " doesn't exist"));
         if(course.getCourseCode() != null){
             existingCourse.setCourseCode(course.getCourseCode());
         }
@@ -54,6 +58,10 @@ public class CourseServicesImpl implements CourseServices {
 
     @Override
     public void removeCourse(Integer id) {
+        if (!repository.existsById(id)) {
+            throw new CourseNotFoundException("Course with id " + id + " doesn't exist.");
+        }
+
         repository.deleteById(id);
     }
 }
